GH-103480: make sysconfig a package
